+++ lib/src/helper/ValidatorHelper.dart
@@ -3,12 +3,12 @@ import 'TextHelper.dart';
 class ValidatorHelper {
   static String errorEmpty = 'Tidak boleh kosong';
 
-  String validateText(String value) {
+  String? validateText(String value) {
     if (value.isEmpty) return errorEmpty;
     return null;
   }
 
-  String validateObject(dynamic value) {
+  String? validateObject(dynamic value) {
     if (value == null) return errorEmpty;
     return null;
   }
@@ -18,13 +18,13 @@ class ValidatorHelper {
   static const maxSkpd = 16;
   static const maxNpwp = 20;
 
-  String validateNop(String value) {
+  String? validateNop(String value) {
     if (value.isEmpty) return errorEmpty;
     if (value.length < maxNop) return 'NOP tidak sesuai';
     return null;
   }
 
-  String validateMoney(String valueWithCurency, int minimalValue) {
+  String? validateMoney(String valueWithCurency, int minimalValue) {
     var value = TextHelper().removeCurrency(valueWithCurency);
     if (value.isEmpty) return errorEmpty;
     try {
@@ -36,29 +36,29 @@ class ValidatorHelper {
     return null;
   }
 
-  String validateYear(String value) {
+  String? validateYear(String value) {
     if (value.isEmpty)
       return errorEmpty;
     else if (value.length != 4) return 'Not valid';
     return null;
   }
 
-  String validateDropDown(dynamic value) {
+  String? validateDropDown(dynamic value) {
     if (value == null) return errorEmpty;
     return null;
   }
 
-  String validateMobile(String value) {
+  String? validateMobile(String value) {
     if (value.isEmpty) return errorEmpty;
     if (value.length < 10) return 'Harus lebih dari 9 angka';
     return null;
   }
 
-  String validateEmail(String value) {
+  String? validateEmail(String value) {
     if (value.isEmpty) return errorEmpty;
     Pattern pattern =
         r'^(([^<>()[\]\\.,;:\s@\"]+(\.[^<>()[\]\\.,;:\s@\"]+)*)|(\".+\"))@((\[[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\])|(([a-zA-Z\-0-9]+\.)+[a-zA-Z]{2,}))$';
-    RegExp regex = new RegExp(pattern);
+    RegExp regex = new RegExp(pattern as String);
     if (!regex.hasMatch(value)) return 'Email tidak valid';
     return null;
   }