+++ lib/src/model/LocalReceivedNotification.dart
@@ -1,6 +1,3 @@
-import 'package:flutter/cupertino.dart';
-import 'package:flutter/material.dart';
-
 class LocalReceivedNotification {
   final int id;
   final String title;
@@ -8,9 +5,9 @@ class LocalReceivedNotification {
   final String payload;
 
   LocalReceivedNotification({
-    @required this.id,
-    @required this.title,
-    @required this.body,
-    @required this.payload,
+    required this.id,
+    required this.title,
+    required this.body,
+    required this.payload,
   });
 }